\name{debrisThickness_measured}
\alias{debrisThickness_measured}
\docType{data}
\title{
Data: Debris thickness (in-situ)
}
\description{
Debris thickness "measurements" at the Liligo Glacier (Karakoram, Pakistan)
}
\usage{data(debrisThickness_measured)}
\format{An object of class \code{'numeric'}.}
\details{
Unit: m

Note: The values of the dataset do not represent real field measurements.
}
\examples{
data(debrisThickness_measured)
print(debrisThickness_measured)
}
\keyword{datasets}
