\name{lst_measured}
\alias{lst_measured}
\docType{data}
\title{
Data: Land surface temperature (in-situ)
}
\description{
Land surface temperature "measurements" at the Liligo Glacier (Karakoram, Pakistan)
}
\usage{data(lst_measured)}
\format{An object of class \code{'numeric'}.}
\details{
Unit: K

Note: The values of the dataset do not represent real field measurements.
}
\examples{
data(lst_measured)
print(lst_measured)
}
\keyword{datasets}
