\name{selectedCoordinates}
\alias{selectedCoordinates}
\docType{data}
\title{
Data: Locations of in-situ measurement
}
\description{
A selection of locations of in-situ "measurements" at the Liligo Glacier (Karakoram, Pakistan)
}
\usage{data(selectedCoordinates)}
\format{An object of class \code{'numeric'}.}
\details{
Coordinates in UTM 43 N
x = longitude
y = latitude

Note: The values of the dataset do not represent real field measurements.
}
\examples{
data(selectedCoordinates)
print(selectedCoordinates)
}
\keyword{datasets}
